+++ webextension/js/variousFeatures/refresh-data-loader.js
@@ -3,10 +3,6 @@ import freshRss from "../platforms/freshrss.js";
 import {WebsiteData} from "./website-data.js";
 import {getPreference} from "../classes/chrome-preferences.js";
 
-export const dataStorageArea = chrome.storage.session ?? chrome.storage.local,
-	refreshDataStorageBase = `_websitesDataStore`
-;
-
 /**
  *
  * @return {Promise<{deviantArt: deviantArt, freshRss: freshRss}>}
@@ -16,24 +12,18 @@ export async function getWebsitesApis() {
 	websites['deviantArt'] = deviantArt;
 	if (!!await getPreference('freshRss_baseUrl')) {
 		websites['freshRss'] = freshRss;
-	} else if (websitesData.has('freshRss')) {
-		websitesData.delete('freshRss');
 	}
 	return websites;
 }
 
-/**
- *
- * @type {Map<string, WebsiteData>}
- */
-export const websitesData = new Map();
 export async function loadStoredWebsitesData() {
-	websitesData.clear();
+	/**
+	 *
+	 * @type {Map<string, WebsiteData>}
+	 */
+	const websitesData = new Map();
 
-	let raw = (await dataStorageArea.get([refreshDataStorageBase])) ?? {};
-	raw = raw[refreshDataStorageBase] ?? {};
-
-	const deviantArtData = !!raw.deviantArt ? WebsiteData.fromJSON(raw.deviantArt) : new WebsiteData();
+	const deviantArtData = new WebsiteData();
 	websitesData.set('deviantArt', deviantArtData);
 	if (!deviantArtData.websiteIcon) {
 		deviantArtData.websiteIcon = deviantArt.defaultFavicon;
@@ -44,7 +34,7 @@ export async function loadStoredWebsitesData() {
 
 	const freshRss_baseUrl = await getPreference('freshRss_baseUrl');
 	if (!!freshRss_baseUrl) {
-		const freshRssData = !!raw.freshRss ? WebsiteData.fromJSON(raw.freshRss) : new WebsiteData();
+		const freshRssData = new WebsiteData();
 		websitesData.set('freshRss', freshRssData);
 		if (!freshRssData.websiteIcon) {
 			freshRssData.websiteIcon = freshRss.defaultFavicon;
@@ -52,8 +42,6 @@ export async function loadStoredWebsitesData() {
 		if (!freshRssData.href) {
 			freshRssData.href = freshRss_baseUrl;
 		}
-	} else if (!!raw.freshRss) { // If no value in 'freshRss_baseUrl' but raw.freshRss data present
-		delete raw.freshRss;
 	}
 
 	return websitesData;